+++ src/commands/stac-sync/__test__/stac.sync.test.ts
@@ -3,9 +3,9 @@ import { beforeEach, describe, it } from 'node:test';
 
 import { fsa } from '@chunkd/fs';
 import { FsMemory } from '@chunkd/source-memory';
-import { createHash } from 'crypto';
 
-import { HashKey, synchroniseFiles } from '../stac.sync.js';
+import { hashBuffer, HashKey } from '../../../utils/hash.js';
+import { synchroniseFiles } from '../stac.sync.js';
 
 describe('stacSync', () => {
   const fs = new FsMemory();
@@ -47,7 +47,7 @@ describe('stacSync', () => {
       JSON.stringify({ title: 'Wellington Collection', description: 'abcd' }),
     );
     const sourceData = await fsa.read('m://source/stac/wellington/collection.json');
-    const sourceHash = '1220' + createHash('sha256').update(sourceData).digest('hex');
+    const sourceHash = hashBuffer(sourceData);
     await fs.write(
       'm://destination/stac/wellington/collection.json',
       JSON.stringify({ title: 'Wellington Collection', description: 'abcd' }),